+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 # Next
 
+### Changed
+
+This release requires Swift 5.3.
+
+Some context: this release heavily relies on where clauses on contextually generic declarations, that are only available in Swift 5.3 - [SE-0267](https://github.com/apple/swift-evolution/blob/master/proposals/0267-where-on-contextually-generic.md).
+
 ### Breaking
 
 * `ViewModelMapping` has been reworked to generic class to capture both View and Model type.
@@ -11,6 +17,8 @@ All notable changes to this project will be documented in this file.
 * `ViewModelMapping` convenience methods `eventsModelMapping` and `eventsViewMapping` have been removed.
 * `mappingCandidates` method now returns `ViewModelMappingProtocol` instead of `ViewModelMapping`
 * `performReaction` methods now accept `IndexPath` for `location` parameter instead of `Any`.
+* Event reactions are moved from extension on `[ViewModelMapping]` to static methods on `EventReaction` due to Swift inability to call methods on sequences that contain protocols.
+* `mappingCandidates` method has been moved to `ViewType` type.
 
 ## [8.1.0](https://github.com/DenTelezhkin/DTModelStorage/releases/tag/8.1.0)